define int64 as long long and uint64 / ThreadId_t as unsigned long long

+++ public/tier0/platform.h
@@ -203,19 +203,19 @@ typedef signed char int8;
 
 #endif // else _WIN32
 
-typedef int16_t		int16;
-typedef uint16_t	uint16;
-typedef int32_t		int32;
-typedef uint32_t	uint32;
-typedef int64_t		int64;
-typedef uint64_t	uint64;
-typedef intptr_t	intp;
-typedef uintptr_t	uintp;
+typedef int16_t				int16;
+typedef uint16_t			uint16;
+typedef int32_t				int32;
+typedef uint32_t			uint32;
+typedef long long			int64;
+typedef unsigned long long	uint64;
+typedef intptr_t			intp;
+typedef uintptr_t			uintp;
 
 #ifdef PLATFORM_64BITS
-typedef uint64_t ThreadId_t;
+typedef unsigned long long 	ThreadId_t;
 #else
-typedef uint32_t ThreadId_t;
+typedef uint32_t 			ThreadId_t;
 #endif
 
 //-----------------------------------------------------------------------------
